# ops-ui/openapi/codex-api.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ListPipelineJobsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PipelineJobSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "ListPipelineJobsResponse",
        "type": "object"
      },
...
      "MicroEventBatchExtractRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "limit": 1,
            "model": "gpt-5.5",
            "overlapMinutes": 5,
            "reasoningEffort": "medium",
            "regenerateSucceeded": false,
            "retryFailed": false,
            "windowMinutes": 30
          }
        ],
        "properties": {
          "limit": {
            "default": 1,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "enum": [
                  "gpt-5.5",
                  "gpt-5.4",
                  "gpt-5.4-mini"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "overlapMinutes": {
            "default": 5,
            "maximum": 239.0,
            "minimum": 0.0,
            "title": "Overlapminutes",
            "type": "integer"
          },
          "reasoningEffort": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "xhigh"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoningeffort"
          },
          "regenerateSucceeded": {
            "default": false,
            "title": "Regeneratesucceeded",
            "type": "boolean"
          },
          "retryFailed": {
            "default": false,
            "title": "Retryfailed",
            "type": "boolean"
          },
          "windowMinutes": {
            "default": 30,
            "maximum": 240.0,
            "minimum": 1.0,
            "title": "Windowminutes",
            "type": "integer"
          }
        },
        "title": "MicroEventBatchExtractRequest",
        "type": "object"
      },
      "MicroEventBatchExtractResponse": {
        "properties": {
          "alreadySatisfiedCount": {
            "title": "Alreadysatisfiedcount",
            "type": "integer"
          },
          "failedCount": {
            "title": "Failedcount",
            "type": "integer"
          },
          "ineligibleCount": {
            "title": "Ineligiblecount",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MicroEventExtractResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "processedCount": {
            "title": "Processedcount",
            "type": "integer"
          },
          "requestedCount": {
            "title": "Requestedcount",
            "type": "integer"
          },
          "scannedCount": {
            "title": "Scannedcount",
            "type": "integer"
          },
          "skippedCount": {
            "title": "Skippedcount",
            "type": "integer"
          },
          "succeededCount": {
            "title": "Succeededcount",
            "type": "integer"
          },
          "timedOutCount": {
            "title": "Timedoutcount",
            "type": "integer"
          }
        },
        "required": [
          "requestedCount",
          "processedCount",
          "succeededCount",
          "failedCount",
          "skippedCount",
          "timedOutCount",
          "scannedCount",
          "alreadySatisfiedCount",
          "ineligibleCount",
          "items"
        ],
        "title": "MicroEventBatchExtractResponse",
        "type": "object"
      },
...
      "MicroEventEnqueueRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "limit": 20,
            "model": "gpt-5.5",
            "overlapMinutes": 5,
            "reasoningEffort": "medium",
            "regenerateSucceeded": false,
            "retryFailed": false,
            "target": "selected_videos",
            "videoIds": [
              1,
              2,
              3
            ],
            "windowMinutes": 30
          }
        ],
        "properties": {
          "channelId": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channelid"
          },
          "limit": {
            "default": 20,
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "enum": [
                  "gpt-5.5",
                  "gpt-5.4",
                  "gpt-5.4-mini"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "overlapMinutes": {
            "default": 5,
            "maximum": 239.0,
            "minimum": 0.0,
            "title": "Overlapminutes",
            "type": "integer"
          },
          "reasoningEffort": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "xhigh"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoningeffort"
          },
          "regenerateSucceeded": {
            "default": false,
            "title": "Regeneratesucceeded",
            "type": "boolean"
          },
          "retryFailed": {
            "default": false,
            "title": "Retryfailed",
            "type": "boolean"
          },
          "search": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search"
          },
          "target": {
            "default": "next_eligible",
            "enum": [
              "selected_videos",
              "current_filters",
              "next_eligible"
            ],
            "title": "Target",
            "type": "string"
          },
          "taskStatus": {
            "anyOf": [
              {
                "enum": [
                  "pending",
                  "running",
                  "succeeded",
                  "failed",
                  "timed_out",
                  "no_transcript",
                  "skipped",
                  "canceled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taskstatus"
          },
          "videoIds": {
            "items": {
              "type": "integer"
            },
            "maxItems": 200,
            "title": "Videoids",
            "type": "array"
          },
          "windowMinutes": {
            "default": 30,
            "maximum": 240.0,
            "minimum": 1.0,
            "title": "Windowminutes",
            "type": "integer"
          }
        },
        "title": "MicroEventEnqueueRequest",
        "type": "object"
      },
      "MicroEventEnqueueResponse": {
        "properties": {
          "alreadyPendingCount": {
            "title": "Alreadypendingcount",
            "type": "integer"
          },
          "alreadyRunningCount": {
            "title": "Alreadyrunningcount",
            "type": "integer"
          },
          "alreadySucceededCount": {
            "title": "Alreadysucceededcount",
            "type": "integer"
          },
          "enqueuedCount": {
            "title": "Enqueuedcount",
            "type": "integer"
          },
          "ineligibleCount": {
            "title": "Ineligiblecount",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MicroEventEnqueueItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "requestedCount": {
            "title": "Requestedcount",
            "type": "integer"
          },
          "scannedCount": {
            "title": "Scannedcount",
            "type": "integer"
          },
          "skippedFailedCount": {
            "title": "Skippedfailedcount",
            "type": "integer"
          }
        },
        "required": [
          "requestedCount",
          "scannedCount",
          "enqueuedCount",
          "alreadyPendingCount",
          "alreadyRunningCount",
          "alreadySucceededCount",
          "skippedFailedCount",
          "ineligibleCount",
          "items"
        ],
        "title": "MicroEventEnqueueResponse",
        "type": "object"
      },
...
      "MicroEventExtractRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "model": "gpt-5.5",
            "overlapMinutes": 5,
            "reasoningEffort": "medium",
            "regenerateSucceeded": false,
            "retryFailed": false,
            "windowMinutes": 30
          }
        ],
        "properties": {
          "model": {
            "anyOf": [
              {
                "enum": [
                  "gpt-5.5",
                  "gpt-5.4",
                  "gpt-5.4-mini"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "overlapMinutes": {
            "default": 5,
            "maximum": 239.0,
            "minimum": 0.0,
            "title": "Overlapminutes",
            "type": "integer"
          },
          "reasoningEffort": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "xhigh"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoningeffort"
          },
          "regenerateSucceeded": {
            "default": false,
            "title": "Regeneratesucceeded",
            "type": "boolean"
          },
          "retryFailed": {
            "default": false,
            "title": "Retryfailed",
            "type": "boolean"
          },
          "windowMinutes": {
            "default": 30,
            "maximum": 240.0,
            "minimum": 1.0,
            "title": "Windowminutes",
            "type": "integer"
          }
        },
        "title": "MicroEventExtractRequest",
        "type": "object"
      },
...
      "PipelineJobDetailResponse": {
        "properties": {
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/PipelineJobAttemptResponse"
            },
            "title": "Attempts",
            "type": "array"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/PipelineChannelOutputResponse"
            },
            "title": "Channels",
            "type": "array"
          },
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "externalApiCalls": {
            "items": {
              "$ref": "#/components/schemas/ExternalApiCallSummaryResponse"
            },
            "title": "Externalapicalls",
            "type": "array"
          },
          "externalKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Externalkey"
          },
          "inputHash": {
            "title": "Inputhash",
            "type": "string"
          },
          "inputJson": {
            "additionalProperties": true,
            "title": "Inputjson",
            "type": "object"
          },
          "jobId": {
            "title": "Jobid",
            "type": "integer"
          },
          "microEventExtractions": {
            "items": {
              "$ref": "#/components/schemas/PipelineMicroEventExtractionOutputResponse"
            },
            "title": "Microeventextractions",
            "type": "array"
          },
          "parentJobId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentjobid"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed",
              "skipped",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          },
          "step": {
            "title": "Step",
            "type": "string"
          },
          "subjectId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectid"
          },
          "subjectType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjecttype"
          },
          "transcriptCues": {
            "items": {
              "$ref": "#/components/schemas/PipelineTranscriptCueOutputResponse"
            },
            "title": "Transcriptcues",
            "type": "array"
          },
          "transcripts": {
            "items": {
              "$ref": "#/components/schemas/PipelineTranscriptOutputResponse"
            },
            "title": "Transcripts",
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "videos": {
            "items": {
              "$ref": "#/components/schemas/PipelineVideoOutputResponse"
            },
            "title": "Videos",
            "type": "array"
          }
        },
        "required": [
          "jobId",
          "step",
          "status",
          "subjectType",
          "subjectId",
          "externalKey",
          "inputJson",
          "inputHash",
          "parentJobId",
          "createdAt",
          "updatedAt",
          "completedAt",
          "attempts",
          "externalApiCalls",
          "channels",
          "videos",
          "transcripts",
          "transcriptCues",
          "microEventExtractions"
        ],
        "title": "PipelineJobDetailResponse",
        "type": "object"
      },
...
      "RetryPipelineJobResponse": {
        "properties": {
          "jobAttemptId": {
            "title": "Jobattemptid",
            "type": "integer"
          },
          "jobId": {
            "title": "Jobid",
            "type": "integer"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed",
              "skipped",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          },
          "step": {
            "title": "Step",
            "type": "string"
          }
        },
        "required": [
          "jobId",
          "jobAttemptId",
          "step",
          "status",
          "result"
        ],
        "title": "RetryPipelineJobResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "REST API wrapper for Codex SDK run/account/logout workflows.",
    "title": "Codex SDK CLI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/pipeline/jobs": {
      "get": {
        "operationId": "list_pipeline_jobs_pipeline_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "step",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Step"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "running",
                    "succeeded",
                    "failed",
                    "skipped",
                    "canceled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "channelId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channelid"
            }
          },
          {
            "in": "query",
            "name": "subjectType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subjecttype"
            }
          },
          {
            "in": "query",
            "name": "subjectId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subjectid"
            }
          },
          {
            "in": "query",
            "name": "externalKey",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Externalkey"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPipelineJobsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pipeline Jobs",
        "tags": [
          "pipeline-jobs"
        ]
      }
    },
    "/pipeline/jobs/{job_id}": {
      "get": {
        "operationId": "get_pipeline_job_pipeline_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineJobDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pipeline Job",
        "tags": [
          "pipeline-jobs"
        ]
      }
    },
    "/pipeline/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_pipeline_job_pipeline_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryPipelineJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Pipeline Job",
        "tags": [
          "pipeline-jobs"
        ]
      }
    },
...
    "/video-tasks/micro-event-extract": {
      "post": {
        "operationId": "extract_all_video_micro_events_video_tasks_micro_event_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MicroEventBatchExtractRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicroEventBatchExtractResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract All Video Micro Events",
        "tags": [
          "micro-events"
        ]
      }
    },
    "/video-tasks/micro-event-extract/enqueue": {
      "post": {
        "operationId": "enqueue_video_micro_events_video_tasks_micro_event_extract_enqueue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MicroEventEnqueueRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicroEventEnqueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enqueue Video Micro Events",
        "tags": [
          "micro-events"
        ]
      }
    },
...
    "/videos/{video_id}/micro-event-extractions/latest": {
      "get": {
        "operationId": "get_latest_video_micro_event_extraction_videos__video_id__micro_event_extractions_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicroEventExtractionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Video Micro Event Extraction",
        "tags": [
          "micro-events"
        ]
      }
    },
    "/videos/{video_id}/micro-event-extractions/{video_task_id}": {
      "get": {
        "operationId": "get_video_micro_event_extraction_videos__video_id__micro_event_extractions__video_task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Video Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "video_task_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Video Task Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicroEventExtractionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Video Micro Event Extraction",
        "tags": [
          "micro-events"
        ]
      }
    },
...
    "/videos/{video_id}/video-tasks/micro-event-extract": {
      "post": {
        "operationId": "extract_video_micro_events_videos__video_id__video_tasks_micro_event_extract_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MicroEventExtractRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicroEventExtractResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract Video Micro Events",
        "tags": [
          "micro-events"
        ]
      }
    },
...
  }
}